Add files via upload
A workshop organized by Coder's Ready. This is day 1 of this workshop in this Data-Science Day-1 workshop learn basic of python and intermediate of python like we learn about conditional statement and control structure. And also learn about how python automatically detect the data type that means it doesn't need to tell the code to what kind of data type we store in variable